crop and soil service update fix duplicate names

+++ src/crop/crop.service.ts
@@ -2,6 +2,7 @@ import {
   Injectable,
   NotFoundException,
   BadRequestException,
+  ConflictException,
 } from "@nestjs/common";
 import { InjectEntityManager, InjectRepository } from "@nestjs/typeorm";
 import { EntityManager, Repository } from "typeorm";
@@ -9,6 +10,7 @@ import { validate } from "class-validator";
 import { Crop } from "./crop.entity";
 import { CreateCropDto } from "./dtos/create-crop.dto";
 import { UpdateCropDto } from "./dtos/update-crop.dto";
+import { GrowingCropPeriod } from "src/growing-crop-period/growing-crop-period.entity";
 
 @Injectable()
 export class CropService {
@@ -39,12 +41,27 @@ export class CropService {
     const { name } = createCropDto;
 
     const existingCrop = await this.cropRepository.findOne({
+      withDeleted: true,
       where: { name },
     });
 
-    if (createCropDto.name === existingCrop.name) {
-      throw new BadRequestException(`Crop already exists!`);
+    if (existingCrop) {
+      // If the soil exists and is soft-deleted, restore it
+      if (existingCrop.deleted) {
+        existingCrop.deleted = null;
+        return await this.cropRepository.save(existingCrop);
+      } else {
+        // If the soil is not soft-deleted, throw a conflict exception
+        throw new ConflictException(`Crop ${name} already exists`);
+      }
     }
+    // const existingCrop = await this.cropRepository.findOne({
+    //   where: { name },
+    // });
+
+    // if (createCropDto.name === existingCrop.name) {
+    //   throw new BadRequestException(`Crop already exists!`);
+    // }
 
     const newCrop = this.cropRepository.create({ name });
     const newCreatedCrop = await this.cropRepository.save(newCrop);
@@ -58,6 +75,17 @@ export class CropService {
     }
 
     if (updateCropDto.name) {
+      const existingCropWithSameName = await this.cropRepository.findOne({
+        where: { name: updateCropDto.name },
+      });
+
+      if (existingCropWithSameName && existingCropWithSameName.id !== id) {
+        throw new ConflictException(
+          `Crop with name ${updateCropDto.name} already exists`,
+        );
+      }
+
+      // Update the crop name
       await this.cropRepository.update(id, { name: updateCropDto.name });
     }
 
@@ -78,16 +106,17 @@ export class CropService {
       throw new NotFoundException(`Crop with id ${id} not found`);
     }
 
-    // const isCropAssociatedWithGrowingCropPeriods =
-    //   await this.growingCropPeriodService.isCropAssociatedWithGrowingCropPeriods(
-    //     id,
-    //   );
+    const isCropAssociatedWithGrowingCropPeriod = await this.entityManager
+      .getRepository(GrowingCropPeriod)
+      .createQueryBuilder("growingCropperiod")
+      .where("growingCropperiod.crop_id = :id", { id })
+      .getCount();
 
-    // if (isCropAssociatedWithGrowingCropPeriods) {
-    //   throw new BadRequestException(
-    //     `This crop with ID ${id} has associated GrowingCropPeriods. Cannot update the crop.`,
-    //   );
-    // }
+    if (isCropAssociatedWithGrowingCropPeriod > 0) {
+      throw new BadRequestException(
+        `This crop with ID ${id} has associated growingCropPeriod. Cannot delete the crop.`,
+      );
+    }
 
     await this.cropRepository.softDelete({ id });
 
@@ -107,16 +136,17 @@ export class CropService {
       throw new NotFoundException(`Crop with id ${id} not found`);
     }
 
-    // const isCropAssociatedWithGrowingCropPeriods =
-    //   await this.growingCropPeriodService.isCropAssociatedWithGrowingCropPeriods(
-    //     id,
-    //   );
+    const isCropAssociatedWithGrowingCropPeriod = await this.entityManager
+      .getRepository(GrowingCropPeriod)
+      .createQueryBuilder("growingCropperiod")
+      .where("growingCropperiod.crop_id = :id", { id })
+      .getCount();
 
-    // if (isCropAssociatedWithGrowingCropPeriods) {
-    //   throw new BadRequestException(
-    //     `This crop with ID ${id} has associated GrowingCropPeriods. Cannot update the crop.`,
-    //   );
-    // }
+    if (isCropAssociatedWithGrowingCropPeriod > 0) {
+      throw new BadRequestException(
+        `This crop with ID ${id} has associated growingCropPeriod. Cannot delete the crop.`,
+      );
+    }
 
     await this.cropRepository.remove(existingCrop);
